+++ generators/view/templates/states/state.default.js
@@ -1,0 +1,25 @@
+/**
+ <%- include('../../../common/partial/header.txt') %>
+ */
+(function (angular) {
+	'use strict';
+
+	angular
+		.module('<%= appNameCamel %>')
+		.config('<%= configStateName %>');
+
+	<%= configStateName %>.$inject = [
+		'$stateProvider'
+	];
+
+	function <%= configStateName %>($stateProvider) {
+		$stateProvider
+			.state('<%= viewNameCamel %>', {
+				url         : '<%= viewUrl %>',
+				templateUrl : 'views/<%= viewNameCamel %>.template.html',
+				controller  : '<%= controllerName %>',
+				controllerAs: 'vm'
+			});
+	}
+
+})(window.angular);