+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.4.1]
 ### Added
 - Docs folder, Code of conduct, issue template and pull request template by merging the master into the develop
+- Abstract state option
 
 ## [0.4.0]
 ### Added